html/app: Add support for ?search={URI}
This allows folks to link to PromeCIeus with a particular job-detail
page in mind, and folks clicking through won't have to bother pasting
in the URI or clicking "Generate".  For example, the job-detail page
itself may want to link PromeCIeus with it's own location in the
search query parameter.

Calling handleSearchInput from the new 'search()' function means that
folks typing into the form field and then clicking "Generate" will get
handleSearchInput called twice, which is not a big deal.  And it means
that folks calling search() from other places can't forget to set the
state.

The new querySearch state ensures we only trigger the auto-search
logic when the query value changes.

searchParams() business is based on [1].

[1]: https://developer.mozilla.org/en-US/docs/Web/API/URL/searchParams#Example