Add custom prometheus build
This adds a new prometheus image based on v2.39.1 that has WAL checksum
checks disabled. This helps when reading tars from GCP jobs, that for
some reason have corrupted WAL records.

Proper snapshotting would be the correct solution, we do not want to
expose the admin API of Prometheus however as it also allows
unauthenticated deletes.