helpers: remove ensure func
This was originally supposed to be a super-quick fail-fast for bad
input, but a) it's touching a Deck/Prow URL which may be super slow
to load and entirely not necessary for the actual functionality and b)
for direct tar links, the HEAD is rejected by the bucket. We parse the
URL directly after this, in `getTarURLFromProw()`, and we will still
fail very quickly so it seems reasonable to remove this HEAD as
superfluous.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>